refactor: move account menu to package header
